Added a WindowController and LineDrawer to properly render the game
A WindowController will control any changes to the window and it's resolution. This will ensure the game is properly drawn to the window. It allows the game to render in 3 versions: True, Stretch and TrueRatioStretch. This script will control their activation, but the lineDrawer controls their effects.

The lineDrawer will be used to render the game and it's components. Currently it will only render it's Line objects using Debug.DrawLine. It does render the lines in different sizes depending on it's resolution mode, whether the lines are made to fit the screen and whether they aim to keep the game area's ratio.