Added a gameController to handle game creation. Line rendering is now done using the graphics class.
A gameController script has been introduced. This script will control how the game is created. So far, it creates the play area's edges and size and sends it over to the line renderer.

The line renderer now renders the lines to the cameras using the graphics.DrawMesh call. Each line now has a reference to their own mesh and vertices.

Also, the Line class is now it's own script with functions that handle mesh creation and such.