The line's vertices now calculate their position relative to the screen size and render mode.
The lines now take into account the window's sizes and a new edgeBufferSize varaible. This allows the lines to properly place themselves in relation to the camera's view. This does mean that each line is now required to be associated to a windowController. Currently, only one windowcontroller will exist as we have just simply put the required variables as static so that we do not need each line to track a reference to it's windowController.

The bug involving the meshes not appearing in the game window seems to be how the unity game window handles resizing. When resized to a given size, sometimes an edge wont render. However, when restarting the game, the edge will be properly rendered.