Added more functionnality to lines and corners
Added point detection for lines. We can now determine if a given point is on a line.

When a corner adds a line to one of it's orthogonal directions, it will also set itself to either the line's start or end corners. If the corner object is not placed onto the line's start or end position, print an error.