Lines now track the players that are attached to them
Each line now tracks the players that have linked themselves to it. This is done by having the player update themselves in the line's linkedPlayers list.

The lines can now detect if a player is moving in a direction along a line that will make them collide with another player. Currently, it just prints to the console when the player moves in a direction that leads into another player before reaching the corner.