Added player collisions on a per-line basis
The players can now collide into eachother if they are on the same line. The collision actually does not have the players touch eachother, but a preset playerSizebuffer value in the ScanAhead function determines how close two players can get before they are unable to approach eachother any further.

Currently, the collision detection of a player within a range is inclusive, but it specifically ignores collisions for players that are on the same game position. That is because if it were to catch collisions of players on the same spot, neither of them would be able to move away from their position.

This new collision detection method also controls player movement, so previous movement functions have been removed.