Further imrpoved the player scanning
When scanning for players when moving a player, detect when the player is approaching a corner and print out the lines attached to said corner to indicate they also need to be scanned for player. A new boolean is introduced to control when a player size buffer is included in the given distances.