Implemented collisions past corners
Players will now scan ahead past corners to check if anothern player is in their way. This implementation uses the PredeterminePlayerMovement() function recursively, so if there is a corner with very short lines and many more corners, this function will take a longer time to run and may not even function properly at all. Further testing in those scenarios are needed.