Added "corner snapping" to the player movement
When the player is inputting a direction that is perpendicular to their current line, the game will search the nearby corners to see if there are any nearby lines that are in the direction of the player's given direction. This is so that when the player is inputting up on a horizontal line BUT they are placed close to a corner that has an attachedLine going up, the game will change the used input to point towards that corner so that they "snap" to the corner to go up the line.