Fixed some bugs involving the line snapping
There were a few bugs fixed that involved the line snapping. When snapping towards a line, if there was another corner along the direction, the GetClosestLineTowards function would be stuck in an infinite loop.

Also, when the player's direction snapped them towards a direction and brough them to a corner, the player would not change their current line to reflect the desired change from one line to the next.

Now, the player can properly utilize the line snapping mechanic when multiple corners are nearby.

Also changed the gameArea around to allow for more testing of the line snapping mechanic