Added a SplitLine() function to the Line object
The SplitLine() function is able to split a line into two, adding a new corner between them. Splitting a line requires a position as to where it will be split. A line cannot be split on it's corner, as there is nothing to split at that point. Currently, it does not handle player reposition if a line is split while a player is on it.