Added player states and state changing. Included basis for grid.
The player can now be in a given state. Each state has seppearete controls for player movement, using UpdatePlayerInputDirection() to handle their primary and secondairy inputs to extract the wanted direction. The MovePlayerRequest() will now handle how the player moves depending on their current state.

A grid value has been added to the gameController. this value will be used to keep the player from drawing on an invisible grid, where each "square" in the grid has a length if gridSize. This value will be used with the player when they want to leave their current line to start drawing.

Also, the new DistanceToCorner() function is also used in PredeterminePlayerMovement(). The previous code used to find the closest corner is commented out as it may be useful later.