Pre-Drawing state will now snap the player to corners and grid marks.
The grid size is now in use. The player, when wanting to draw a line, will hold down their first button key. For player1, it defaults to spacebar. When holding this button, movement will be focused only on the primary input. When inputting a direction perpendicular of the current line, the player will use DistanceToClosestGrid() to find where the nearest "drawing point" and move towards it.

Currently, moving in the predrawing state does not take into account other players. A fix involving the use of PredeterminePlayerMovement() is an ideal solution.