The Pre-drawing state now creates a new line for the player to use to draw
The Pre-Drawing state will now split the player's current line if needed so that the player can enter the drawing state while on a corner. When in the drawing state, the player will need to be on their own line. They will be moving a corner around the game area. This corner and line should be a different type: a temporary corner/line possibly. The main issue involves the corner being created on the same position as another corner.

This has added a few changes to how the gameController and lineDrawer handle their line and corner objects. The gameController adds each line and corner to the drawer individually. Later, we end up removing these lists and simply use a direct reference to the gameController.

Also, moving in the PreDrawing state now uses PredeterminePlayerMovement to prevent moving past players.