Inputting sideways directions while drawing will snap the player to the grid mark
When in the drawing state and inputting a direction perpendicular to the current line, the player will look ahead to the nearest grid mark and snap to it. When trying to move perpendicularly on a line while on the grid mark, the player should begin drawing another line at that point.