Player can now draw perpendicular lines in the drawing state
The player can now change their direction in the drawing state and can now draw multiple lines. When creating a line while drawing, it is created as a point (start = end), then sent off to the gameController. The gameController and lineDrawer is able to treat points by not adding any extra size to the "line", making it invisible. This is so that we send the line to drawn as a point, and right after we set it's new end position so that the functions can treat it as a line.

An issue that can pressumably occur is when we are adding a line, we send it as a point then change it's end position so it's treated as a line. This is done during the drawing state and when changing lines. If we change line but don't traverse it so that it's size changes, this would mean an error has occured where we tried to progress along the line but had no remaining distance to travel. This is important because if we create the point line then change the window size, the line will be re-created with it's old end point that we set only for the program to handle and not to draw. This will however only be a potential error that is caught on line 316 by an throw exception.

Also, we are currently able to pass through lines and create corners on positions that already contains corners. This is obviously not intended as we want the player to not be able to draw ontop of other lines.