Added a line collision function
We can now detect collisions between lines. The detection uses the fact that all lines are stuck onto one axis, as it makes certain assumptions that reduce the required time to calculate a collision. Currently, we are printing whether the player's current line collides with the first line in the gameController's line list.