Drawing state now checks for line collisions. Updated collision checks.
The drawing state will now check if a collision will occur between the player's upcomming drawing and any pre-existant lines in the gameController.

One of the collision checks was also incorrect as it was getting false positives on lines that are parallel.

Also, when moving in the drawing state and colliding into a line, the collided lines are drawn using Debug.DrawLine. This i